+++ CHANGELOG.md
@@ -9,11 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- detailed run output on `--verbosity=3` (#748)
+- detailed run output on `run --verbosity=3` (#748)
+- detailed run output on `test --verbosity=3` (#755)
 
 ### Changed
 ### Deprecated
 ### Removed
+
+- remove `foldr` (#760)
+
 ### Fixed
 
 - Effects no longer break when applying constant operators (#759)